feat: user profile screen widget test
